# 默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "订单主流程接口（trade 模块）",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/forum/post/create": {
      "post": {
        "summary": "发布帖子",
        "deprecated": false,
        "description": "创建新的论坛帖子",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostCreateDTO"
              },
              "example": {
                "title": "这是一个测试帖子标题",
                "content": "这是帖子的正文内容，可以包含富文本内容，如图片、链接等。",
                "categoryId": 1,
                "tagIds": [
                  1,
                  2
                ],
                "postType": "normal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/forum/post/update": {
      "put": {
        "summary": "编辑帖子",
        "deprecated": false,
        "description": "更新帖子信息",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/list": {
      "post": {
        "summary": "分页查询帖子列表",
        "deprecated": false,
        "description": "获取帖子列表，支持分页和搜索",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSearchDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/{postId}": {
      "get": {
        "summary": "获取帖子详情",
        "deprecated": false,
        "description": "根据帖子ID获取帖子详情",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除帖子",
        "deprecated": false,
        "description": "删除帖子",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/{postId}/like": {
      "post": {
        "summary": "点赞/取消点赞帖子",
        "deprecated": false,
        "description": "对帖子进行点赞或取消点赞",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/{postId}/collect": {
      "post": {
        "summary": "收藏/取消收藏帖子",
        "deprecated": false,
        "description": "对帖子进行收藏或取消收藏",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/{postId}/share": {
      "post": {
        "summary": "转发帖子",
        "deprecated": false,
        "description": "分享帖子到指定渠道",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分享成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/user/{authorId}": {
      "get": {
        "summary": "查询用户发布的帖子",
        "deprecated": false,
        "description": "获取指定用户发布的帖子列表",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/admin/audit/{postId}": {
      "put": {
        "summary": "审核帖子（管理员）",
        "deprecated": false,
        "description": "审核帖子并设置审核结果",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "approved",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "rejectReason",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "审核成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/admin/top/{postId}": {
      "put": {
        "summary": "置顶/取消置顶帖子（管理员）",
        "deprecated": false,
        "description": "设置帖子是否置顶",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "top",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/forum/post/admin/feature/{postId}": {
      "put": {
        "summary": "设为/取消精华帖（管理员）",
        "deprecated": false,
        "description": "设置帖子是否为精华帖",
        "tags": [
          "论坛/论坛帖子管理",
          "论坛帖子管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "featured",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/forum/comment/post/{postId}": {
      "get": {
        "summary": "获取帖子的评论列表",
        "deprecated": false,
        "description": "获取指定帖子的评论列表",
        "tags": [
          "论坛/论坛评论管理",
          "论坛评论管理"
        ],
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        }
      },
      "PostCreateDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "帖子标题"
          },
          "content": {
            "type": "string",
            "description": "帖子内容"
          },
          "categoryId": {
            "type": "integer",
            "description": "分类ID"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "标签ID列表"
          },
          "mediaList": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "description": "媒体类型"
                },
                "url": {
                  "type": "string",
                  "description": "媒体URL"
                }
              }
            },
            "description": "媒体列表"
          }
        }
      },
      "PostUpdateDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "帖子ID"
          },
          "title": {
            "type": "string",
            "description": "帖子标题"
          },
          "content": {
            "type": "string",
            "description": "帖子内容"
          },
          "categoryId": {
            "type": "integer",
            "description": "分类ID"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "标签ID列表"
          },
          "mediaList": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "description": "媒体类型"
                },
                "url": {
                  "type": "string",
                  "description": "媒体URL"
                }
              }
            },
            "description": "媒体列表"
          }
        }
      },
...
      "PostSearchDTO": {
        "type": "object",
        "properties": {
          "pageNum": {
            "type": "integer",
            "description": "页码",
            "default": 1
          },
          "pageSize": {
            "type": "integer",
            "description": "每页数量",
            "default": 10
          },
          "categoryId": {
            "type": "integer",
            "description": "分类ID"
          },
          "tagId": {
            "type": "integer",
            "description": "标签ID"
          },
          "keyword": {
            "type": "string",
            "description": "搜索关键词"
          },
          "sortBy": {
            "type": "string",
            "description": "排序方式",
            "enum": [
              "newest",
              "hot",
              "comments"
            ]
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "servers": [],
  "security": []
}